support for loading and previewing 15+ new sample formats. ld0ds patch: added distortion on ld fungus
missed a buzz-flag, wave import leak/bug

fixed bmx wave decompressor bug

remove unused plugin flags from plugin.h, fix buzz2zzub handling of MIF_MONO_TO_STEREO

added new arguments in zzub_player_load_bmx(): flags for skipping patterns/sequences/waves, and x,y for repositioning the song, fixed problems with wrapper-specific sub-commands in buzz2zzub

buzz2zzub: updated MachineInterface.h to 29-Jun-2008 18:56 + modifications, filter machine flags, abuse the host_ptr if host-version matches buze 0.5.3

added/implemented zzub_player_set_host_info(), host::get_host_info(), host::get_pattern(), host::get_pattern_name(), host::get_pattern_length(), plugin::play_pattern()

ld0ds patches: fungus envelope refresh, midicc enhancements

added a do_event argument on flush_operations(), fix pattern editing weirdness, added two midi plugins from ld0d: miditime and midicc

replace algorithm to determine plugin process order

obey send_events parameter in op_sequencer_set_event::finish()

fix broken midi sync, make new work order on disconnect

ccm save crash fix

ld0ds patch: fix notedelay in fungus

test for wavelevel in zzub_wave_save_sample_range

added lunar port of mda epiano

added ld fungus

simplified the file recorder

simplified the wavetable recorder, added attributes on the built-in recorders for 16, 24 and 32 bit, and continous recording, sample converters also support f32->si24 and si32->si24, minor cleanups

fix plugin/preset saving on vc2005

+++ src/libzzub/wavetable.h
@@ -138,7 +138,7 @@ struct wave_info_ex : wave_info {
 
 	inline unsigned int get_unextended_samples(int level, int samples) {
 		int channels = get_stereo()?2:1;
-		return ceil((samples * get_bytes_per_sample(level)) / 2.0f) + (4/channels);
+		return (unsigned int)ceil((samples * get_bytes_per_sample(level)) / 2.0f) + (4/channels);
 	}
 
 	unsigned int get_sample_count(int level) {
